+++ hw_4_logs_and_test/src/main/java/ua/com/alevel/facade/impl/HallFacadeImpl.java
@@ -5,37 +5,82 @@ import ua.com.alevel.dto.hall.HallUpdateDto;
 import ua.com.alevel.entity.Hall;
 import ua.com.alevel.facade.HallFacade;
 import ua.com.alevel.service.HallService;
+import ua.com.alevel.service.SessionService;
 import ua.com.alevel.service.impl.HallServiceImpl;
+import ua.com.alevel.service.impl.SessionServiceImpl;
+import ua.com.alevel.util.ConstGlobal;
+
+import java.io.IOException;
 
 public class HallFacadeImpl implements HallFacade {
 
     private final HallService hallService = new HallServiceImpl();
+    private final SessionService sessionService = new SessionServiceImpl();
 
     @Override
     public void create(HallCreateDto hallCreateDto) {
+
         Hall hall = createHall(hallCreateDto);
         hallService.create(hall);
     }
 
     @Override
-    public void update(HallUpdateDto hallUpdateDto) {
-        Hall hall = updateHall(hallUpdateDto);
-        hallService.update(hall);
+    public void update(HallUpdateDto hallUpdateDto) throws IOException {
+        if (!exists(hallUpdateDto.getId())) {
+            System.out.println(ConstGlobal.settings.getString("hall.id.not.found") + " " + hallUpdateDto.getId());
+            ConstGlobal.loggerWarn.warn(ConstGlobal.settings.getString("hall.id.not.found") + " " + hallUpdateDto.getId());
+            throw new IOException();
+        } else {
+            Hall hall = updateHall(hallUpdateDto);
+            hallService.update(hall);
+        }
     }
 
     @Override
-    public void delete(Long id) {
-        hallService.delete(id);
+    public void delete(Long id) throws IOException {
+        if (existsHallInSession(id)) {
+            System.out.println(ConstGlobal.settings.getString("hall.used.session"));
+            ConstGlobal.loggerWarn.warn(ConstGlobal.settings.getString("hall.not.delete") + id);
+            throw new IOException();
+        } else if (!exists(id)) {
+            System.out.println(ConstGlobal.settings.getString("hall.id.not.found") + " " + id);
+            ConstGlobal.loggerWarn.warn(ConstGlobal.settings.getString("hall.id.not.found") + " " + id);
+            throw new IOException();
+        } else
+            hallService.delete(id);
     }
 
     @Override
-    public Hall findById(Long id) {
-        return hallService.findById(id);
+    public Hall findById(Long id) throws IOException {
+        if (!exists(id)) {
+            System.out.println(ConstGlobal.settings.getString("hall.id.not.found") + " " + id);
+            ConstGlobal.loggerWarn.warn(ConstGlobal.settings.getString("hall.id.not.found") + " " + id);
+            throw new IOException();
+        } else {
+            ConstGlobal.loggerInfo.info(ConstGlobal.settings.getString("hall.find.by.id") + id);
+            return hallService.findById(id);
+        }
     }
 
     @Override
-    public Hall[] findAll() {
-        return hallService.findAll();
+    public Hall[] findAll() throws IOException {
+        if (hallService.count() == 0) {
+            System.out.println(ConstGlobal.settings.getString("hall.does.not.exist"));
+            ConstGlobal.loggerWarn.warn(ConstGlobal.settings.getString("hall.does.not.exist"));
+            throw new IOException();
+        } else {
+            return hallService.findAll();
+        }
+    }
+
+    @Override
+    public boolean exists(Long id) {
+        return hallService.exists(id);
+    }
+
+    @Override
+    public boolean existsHallInSession(Long id) {
+        return sessionService.existsHallInSession(id);
     }
 
     private Hall createHall(HallCreateDto hallCreate) {
